Adding the service shutdown to the upgrade
Grenade script has two parameters to run it in two parts:
-b Will run the part of grenade that installs the base Devstack
-t Will run the part of Grenade that upgrades the base Devstack to
the target release.
As part of the last steps of the base section, Grenade stops all
the OpenStack services and then run the pre-upgrade resource
verification. This patch moves those steps to the beginning of the
upgrade section so we can install the compute node, since we need
the services to be up in order for the installation to succeed.